#2190 -  POC; Simplification to cloud-ready (true/false) and consolidated into existing OSGi Component
